gugu

+++ src/main/java/org/wa55death405/tp1/repositories/EtudiantRepository.java
@@ -3,6 +3,5 @@ package org.wa55death405.tp1.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.wa55death405.tp1.entities.Etudiant;
 
-public interface EtudiantRepository extends JpaRepository<Etudiant, Long> {
-
+public interface EtudiantRepository extends JpaRepository<Etudiant, Integer> {
 }
